Restructure Form Fields folders
